Fixed a bug : Route to other components only on auth state change & vice-versa
