+++ index.html
@@ -5,7 +5,7 @@
     <link rel="icon" type="image/svg+xml" href="/netflix.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="/src/index.css" />
-    <title>NetflixGPT</title>
+    <title>NetGPT</title>
   </head>
   <body>
     <div id="root"></div>
