+++ src/components/Body.jsx
@@ -1,15 +1,9 @@
-/* eslint-disable react-hooks/exhaustive-deps */
-import React, { useEffect } from "react";
+import React from "react";
 import Login from "./Login";
 import Browse from "./Browse";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
-import { auth } from "../../utils/firebase";
-import { onAuthStateChanged } from "firebase/auth";
-import { useDispatch } from "react-redux";
-import { addUser, removeUser } from "../../utils/userSlice";
 
 const Body = () => {
-  const dispatch = useDispatch();
   const appRouter = createBrowserRouter([
     {
       path: "/",
@@ -20,22 +14,6 @@ const Body = () => {
       element: <Browse />,
     },
   ]);
-
-  useEffect(() => {
-    onAuthStateChanged(auth, (user) => {
-      if (user) {
-        console.log(user);
-        const { uid, email } = user;
-        dispatch(addUser({ uid: uid, email: email }));
-
-        // ...
-      } else {
-        // User is signed out
-        dispatch(removeUser());
-        // ...
-      }
-    });
-  }, []);
 
   return (
     <div>
